+++ Assets/Scripts/LevelComponenets/turret.cs
@@ -6,25 +6,37 @@ public class turret : MonoBehaviour {
     public GameObject Bullet_Emitter;
     public GameObject Bullet;
 
+    public GameObject turretGun;
+    public bool activate;
+
     // Use this for initialization
     void Start() {
+
+        activate = false;
         InvokeRepeating("FireBullets", 0, 2.0f);
         
     }
 
-    // Update is called once per frame
-    void Update()
-    {
-
-    }
-
         void FireBullets()
     {
-        if (turretActivator.activate == true)
+        if (activate == true)
             {
             Instantiate(Bullet, Bullet_Emitter.transform.position, Bullet_Emitter.transform.rotation);
             }
         }
-    
 
+    void OnTriggerEnter(Collider col)
+    {
+
+        if (col.gameObject.tag == "Player")
+        {
+            activate = true;
+        }
+
+    }
+
+    void OnTriggerExit(Collider col)
+    {
+        activate = false;
+    }
 }
