--- Assets/Scripts/LevelComponenets/turretActivator.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: dc3d7a77f87b38244a5cb755f36441ca
-timeCreated: 1475187107
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
